Move C# 8 test cases to separate file
